+++ 3-enrich/scala-common-enrich/src/main/scala/com.snowplowanalytics.snowplow.enrich/common/enrichments/ClientEnrichments.scala
@@ -20,17 +20,13 @@ import java.lang.{Integer => JInteger}
 import scalaz._
 import Scalaz._
 
-// UserAgentUtils
-import eu.bitwalker.useragentutils._
-
 /**
  * Contains enrichments related to the
  * client - where the client is the
  * software which is using the SnowPlow
  * tracker.
  *
- * Enrichments relate to the useragent,
- * browser resolution, etc.
+ * Enrichments relate to browser resolution
  */
 object ClientEnrichments {
   
@@ -42,30 +38,6 @@ object ClientEnrichments {
    * https://github.com/snowplow/snowplow/wiki/snowplow-tracker-protocol#wiki-browserandos
    */
   private val ResRegex = """(\d+)x(\d+)""".r
-
-  /**
-   * Case class to wrap everything we
-   * can extract from the useragent
-   * using UserAgentUtils.
-   *
-   * TODO: update this definition when
-   * we swap out UserAgentUtils for
-   * ua-parser
-   */
-  case class ClientAttributes(
-      // Browser
-      browserName: String,
-      browserFamily: String,
-      browserVersion: Option[String],
-      browserType: String,
-      browserRenderEngine: String,
-      // OS the browser is running on
-      osName: String,
-      osFamily: String,
-      osManufacturer: String,
-      // Hardware the OS is running on
-      deviceType: String,
-      deviceIsMobile: Boolean)
 
   /**
    * Extracts view dimensions (e.g. screen resolution,
@@ -93,43 +65,4 @@ object ClientEnrichments {
       case _ => "Field [%s]: [%s] does not contain valid view dimensions".format(field, res).fail
     }
 
-  /**
-   * Extracts the client attributes
-   * from a useragent string, using
-   * UserAgentUtils.
-   *
-   * TODO: rewrite this when we swap
-   * out UserAgentUtils for ua-parser
-   *
-   * @param useragent The useragent
-   *        String to extract from.
-   *        Should be encoded (i.e.
-   *        not previously decoded).
-   * @return the ClientAttributes or
-   *         the message of the
-   *         exception, boxed in a
-   *         Scalaz Validation
-   */
-  def extractClientAttributes(useragent: String): Validation[String, ClientAttributes] = 
-
-    try {
-      val ua = UserAgent.parseUserAgentString(useragent)
-      val b  = ua.getBrowser
-      val v  = Option(ua.getBrowserVersion)
-      val os = ua.getOperatingSystem
-
-      ClientAttributes(
-        browserName = b.getName,
-        browserFamily = b.getGroup.getName,
-        browserVersion = v map { _.getVersion },
-        browserType = b.getBrowserType.getName,
-        browserRenderEngine = b.getRenderingEngine.toString,
-        osName = os.getName,
-        osFamily = os.getGroup.getName,
-        osManufacturer = os.getManufacturer.getName,
-        deviceType = os.getDeviceType.getName,
-        deviceIsMobile = os.isMobileDevice).success
-    } catch {
-      case e => "Exception parsing useragent [%s]: [%s]".format(useragent, e.getMessage).fail
-    }
 }
